requirements.txt
Flask==2.1.2
Jinja2==3.1.2
nltk==3.7
numpy==1.23.1
scikit-learn==1.1.1
sklearn==0.0
gunicorn==20.1.0
Werkzeug==2.1.2
pdfminer.six==20220524